Revisión procesado ficheros con TIKA, centralización acceso a BBDD, aplicación monitorización actividad y adaptación a mejoras propuestas por pylint.

+++ src/tika_processor.py
@@ -5,11 +5,12 @@ import time
 import shutil
 import subprocess
 import json
-import sqlite3
-from datetime import datetime
 import requests
 from watchdog.observers import Observer
 from watchdog.events import FileSystemEventHandler
+
+# Importar funciones centralizadas desde db_utils.py
+from db_utils import connect_to_db, check_existing_fichero, add_fichero_record
 
 # Directorios
 BASE_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
@@ -33,17 +34,6 @@ ACCEPT_FORMAT = "text/plain"  # Cambia valor según formato ("application/json" 
 DB_PATH = os.path.abspath(
     os.path.join(os.path.dirname(__file__), "../data/sistema_conocimiento.db")
 )
-
-def connect_to_db():
-    """
-    Establece una conexión con la base de datos SQLite.
-
-    :return: Objeto de conexión a la base de datos si existe, de lo contrario None.
-    """
-    if not os.path.exists(DB_PATH):
-        print(f"❌ No se encontró la base de datos en: {DB_PATH}")
-        return None
-    return sqlite3.connect(DB_PATH)
 
 # Comando para iniciar el servidor Tika
 def start_tika_server():
@@ -87,12 +77,13 @@ def process_document(file_path, accept_format):
             return
         else:
             # Eliminar el registro anterior
-            db_conn = sqlite3.connect(DB_PATH)
-            cursor = db_conn.cursor()
-            cursor.execute("DELETE FROM Ficheros WHERE Id = ?", (existing_id,))
-            db_conn.commit()
-            db_conn.close()
-            print(f"🗑️ Registro anterior eliminado para el archivo: {file_name}")
+            db_conn = connect_to_db()
+            if db_conn:
+                cursor = db_conn.cursor()
+                cursor.execute("DELETE FROM Ficheros WHERE Id = ?", (existing_id,))
+                db_conn.commit()
+                db_conn.close()
+                print(f"🗑️ Registro anterior eliminado para el archivo: {file_name}")
 
     # Procesar el archivo con Apache Tika
     start_time = time.time()
@@ -131,9 +122,9 @@ def process_document(file_path, accept_format):
                         content_html = tika_response.get("X-TIKA:content", "")
                         if content_html.strip():
                             output_html_file = os.path.join(
-                                                    PROCESSED_DIR,
-                                                    f"{base_name}_{metodo_extraccion}_Content.html"
-                                                    )
+                                PROCESSED_DIR,
+                                f"{base_name}_{metodo_extraccion}_Content.html"
+                            )
                             with open(output_html_file, "w", encoding="utf-8") as f:
                                 f.write(content_html)
                             print(f"✅ Contenido HTML extraído guardado como: {output_html_file}")
@@ -143,14 +134,14 @@ def process_document(file_path, accept_format):
                             print(
                                 "⚠️ No se encontró contenido en 'X-TIKA:content' "
                                 f"para el archivo: {file_name}"
-                                )
+                            )
                             return
                     elif accept_format == "text/plain":
                         # Guardar el contenido como archivo de texto
                         output_txt_file = os.path.join(
-                                                PROCESSED_DIR,
-                                                f"{base_name}_{metodo_extraccion}_Content.txt"
-                                                )
+                            PROCESSED_DIR,
+                            f"{base_name}_{metodo_extraccion}_Content.txt"
+                        )
                         with open(output_txt_file, "w", encoding="utf-8") as f:
                             f.write(response.text)
                         print(f"✅ Contenido de texto extraído guardado como: {output_txt_file}")
@@ -173,9 +164,9 @@ def process_document(file_path, accept_format):
                     )
                 else:
                     print(
-                            "⚠️ No se pudo extraer texto del archivo: {file_name}. "
-                            f"Código de estado: {response.status_code}"
-                        )
+                        f"⚠️ No se pudo extraer texto del archivo: {file_name}. "
+                        f"Código de estado: {response.status_code}"
+                    )
                 break  # Salir del bucle si se procesa correctamente
             except PermissionError:
                 print(f"⚠️ Archivo en uso, reintentando: {file_name}")
@@ -192,76 +183,6 @@ def sanitize_filename(value):
     return value.replace("/", "_").replace("\\", "_").replace(":", "_").replace("*", "_") \
                 .replace("?", "_").replace("\"", "_").replace("<", "_").replace(">", "_") \
                 .replace("|", "_")
-
-def add_fichero_record(
-        nombre_original, tipo_original, metodo_extraccion, fichero_generado, tipo_extraccion,
-        tiempo_extraccion, observaciones=None
-        ):
-    """
-    Añade un registro a la tabla Ficheros de la base de datos.
-
-    :param DB_PATH: Ruta a la base de datos SQLite.
-    :param nombre_original: Nombre original del archivo.
-    :param tipo_original: Extensión del archivo original.
-    :param metodo_extraccion: Método de extracción utilizado.
-    :param fichero_generado: Ruta del fichero generado tras la extracción.
-    :param tipo_extraccion: Tipo de extracción (formato solicitado).
-    :param tiempo_extraccion: Tiempo que tomó la extracción en segundos.
-    :param observaciones: Observaciones adicionales (opcional).
-    """
-    db_conn = sqlite3.connect(DB_PATH)
-    cursor = db_conn.cursor()
-
-    try:
-        # Extraer solo el nombre del fichero generado (sin la ruta)
-        fichero_generado_nombre = os.path.basename(fichero_generado)
-
-        fecha_extraccion = int(datetime.now().timestamp())
-        cursor.execute("""
-            INSERT INTO Ficheros (
-                nombreOriginal, tipoOriginal, metodoExtraccion, ficheroGenerado, 
-                tipoExtraccion, tiempoExtraccion, observaciones, fechaExtraccion
-            ) VALUES (?, ?, ?, ?, ?, ?, ?, ?)
-        """, (
-            nombre_original, tipo_original, metodo_extraccion, fichero_generado_nombre,
-            tipo_extraccion, tiempo_extraccion, observaciones, fecha_extraccion
-        ))
-        db_conn.commit()
-        print(f"✅ Registro añadido a la base de datos para el archivo: {nombre_original}")
-    except Exception as e:
-        print(f"❌ Error al añadir el registro a la base de datos: {e}")
-    finally:
-        db_conn.close()
-
-def check_existing_fichero(nombre_original, tipo_original, metodo_extraccion):
-    """
-    Comprueba si ya existe un fichero con el mismo nombre, tipo original y método de extracción.
-
-    :param DB_PATH: Ruta a la base de datos SQLite.
-    :param nombre_original: Nombre original del archivo.
-    :param tipo_original: Extensión del archivo original.
-    :param metodo_extraccion: Método de extracción utilizado.
-    :return: El ID del registro existente si se encuentra, de lo contrario None.
-    """
-    db_conn = sqlite3.connect(DB_PATH)
-    cursor = db_conn.cursor()
-
-    try:
-        query = """
-           SELECT Id FROM Ficheros
-           WHERE nombreOriginal = ? AND tipoOriginal = ? AND metodoExtraccion = ?
-        """
-        params = (nombre_original, tipo_original, metodo_extraccion)
-        print("🔍 Ejecutando consulta SQL: {query.strip()} "
-              f"con los parametros: {params}")  # Traza de la consulta
-        cursor.execute(query, params)
-        result = cursor.fetchone()
-        return result[0] if result else None
-    except Exception as e:
-        print(f"❌ Error al comprobar el registro en la base de datos: {e}")
-        return None
-    finally:
-        db_conn.close()
 
 # Monitor de la carpeta de entrada
 class WatcherHandler(FileSystemEventHandler):
@@ -318,4 +239,3 @@ if __name__ == "__main__":
             tika_process.terminate()  # Detener el proceso de Tika
         observer.stop()
     observer.join()
-
